#13067 - controller url변경 및 MetricManualEntity 키값 추가, metricValue 타입 Double변경

+++ src/main/kotlin/co/brainz/itsm/sla/metricManual/dto/MetricManualDto.kt
@@ -13,7 +13,7 @@ data class MetricManualDto(
     val metricId: String = "",
     val metricName: String? = null,
     val referenceDt: LocalDateTime? = null,
-    val metricValue: Long? = null,
+    val metricValue: Double? = null,
     val metricUnitName: String? = null,
     val createDt: LocalDateTime? = null,
     val createUserName: String? = null
